Create a function that concatenates at most n bytes of src string to dest string

+++ 0x06-pointers_arrays_strings/main.h
@@ -1,3 +1,3 @@
 int _putchar(char c);
 char *_strcat(char *dest, char *src);
-char *_strncat(char *dest, char *src, size_t n);
+char *_strncat(char *dest, char *src, int n);
